cleaned testing files
