--- Tests/test_stuff.py
@@ -1,12 +1,0 @@
-import pytest
-import random
-
-
-def test_maybe_pass():
-    assert random.random() > 0.5
-
-def test_pass():
-    assert True
-
-def test_fail():
-    assert False
